sidecar/requirements.txt
flask>=3.0
music21>=9.1
mido>=1.3
gunicorn>=22.0
Pillow>=10.0
